gemini-poc/requirements.txt
# Gemini LLM POC - Requirements

# Google AI (Gemini) SDK
google-generativeai==0.3.2

# Basic dependencies
python-dotenv==1.0.0
requests==2.31.0

# For async operations (optional)
asyncio-extras==1.3.2

# For pretty output
rich==13.7.0